Move keypair details modal to separate folder
