+++ src/components/Cluster/ClusterDetail/KeypairDetailsModal/KeyPairDetailsModal.js
@@ -4,8 +4,8 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import Copyable from 'shared/Copyable';
 
-import GenericModal from '../../Modals/GenericModal';
-import CertificateOrgsLabel from './CertificateOrgsLabel';
+import GenericModal from '../../../Modals/GenericModal';
+import CertificateOrgsLabel from '../CertificateOrgsLabel';
 
 class KeyPairDetailsModal extends React.Component {
   static createDate(date) {
